@@ -17,7 +17,6 @@ classifiers =
     Framework :: Twisted
     Programming Language :: Python
     Programming Language :: Python :: 3
-    Programming Language :: Python :: 3.5
     Programming Language :: Python :: 3.6
     Programming Language :: Python :: 3.7
     Programming Language :: Python :: 3.8
@@ -26,10 +25,10 @@ classifiers =
     Topic :: System :: Systems Administration :: Authentication/Directory :: LDAP
 
 [options]
-python_requires = ~=3.5
+python_requires = ~=3.6
 install_requires =
     passlib
-    Twisted[tls]>=15.5.0
+    Twisted[tls]>=17.9.0
     pyparsing
     six >= 1.7
     zope.interface
